* Added support for new ocarina (Brawl+) * Fixed small issue with invalid amount of partitions * Added support for WIP patches * Added support for BCA codes * Fixed cheat downloading * Changed cheat window (now only shows 4 items on a page) * Moved default location of cheat files * Location of cheat files, wip patches, bca codes and text codes is configurable in wiiflow.ini
Note: this version will look for your cheat codes (gct files and txt files) in another location. You will need to manually move these to the correct locations. These locations should be in the directory that also contains the subdirectories for boxcovers, cache, themes, etc. and are called codes (for gct, bca and wip files) and txtcodes (for txt files).