*Dont shutdown the usb device and start it up before we try to mount it, just try and start it. *Add a function to start the drive up sooner.
*Moved some png/ttf/wav files from the resources dir to data dir and removed the .c and .h files for them.
*Removed resources dir.

NOTE: we need the png's for the dvdskin.c, dvdskin_red.c, and dvdskin_black.c so we can move those to the wffilow/data dir and remove wiiflow/source/data/ dir.

If we would rather use .c and .h files for the png/wav/ttf files, ALL of those files should be moved to a resource folder and all of their .c and .h files need to be included in source/data.  Doing half one way and half the other way just seems unorganized to me.